Update snakemake-relevant-python.qmd
Minor edits.  Changed "drop you hack into R" to "drop you back into R" in section 12.1 and changed path to units.tsv in section 12.2.13 to be "data/config/units.tsv" not "../data/config/units.tsv" since the R project for this repository is on the same level as data, so the path "../data/config/units.tsv" did not exist when I tried running that bit of code.